+++ benchmarking/drugcomb/data_manager.py
@@ -1,0 +1,19 @@
+from pyrelational.data_managers import DataManager
+from pyrelational.datasets import DrugCombDataset
+
+
+class DrugCombDataManager(DataManager):
+
+    def __init__(self, seed: int, batch_size: int = 32):
+
+        dataset = DrugCombDataset()
+        train_indices, test_indices = dataset.data_splits[0]
+        super().__init__(
+            dataset,
+            train_indices=train_indices,
+            test_indices=test_indices,
+            random_label_size=256,
+            hit_ratio_at=5,
+            random_seed=seed,
+            loader_batch_size=batch_size,
+        )
